+++ src/drivers/flashmem/posix.cpp
@@ -100,12 +100,12 @@ const char* getFilename()
     return "flash.bin";
 }
 
-uint32_t getTotalSize()
+uint64_t getTotalSize()
 {
     return FLASH_MEMORY_SIZE;
 }
 
-bool readData(uint8_t* buf, uint32_t address, size_t length)
+bool readData(uint8_t* buf, uint64_t address, size_t length)
 {
     bool result = false;
