+++ src/drivers/flashmem/w25qxx.cpp
@@ -233,12 +233,12 @@ StorageStatistics getStatistics(void)
     return stats;
 }
 
-uint32_t getTotalSize()
+uint64_t getTotalSize()
 {
     return cfg->block_count * BLOCK_SIZE;
 }
 
-bool readData(uint8_t* buf, uint32_t address, size_t length)
+bool readData(uint8_t* buf, uint64_t address, size_t length)
 {
     return readIntoBuffer(address, buf, length);
 }
